requirements.txt
gradio
numpy
scipy
uuid
faster-whisper
openai
python-dotenv
requests
piper-tts
fastrtc
